Add HTML table converter module
